pbs-215--refactor: inqName 통합
