+++ backend/src/main/java/com/pobluesky/backend/domain/lineitem/entity/type/InqName.java
@@ -1,4 +1,4 @@
-package com.pobluesky.backend.domain.lineitem.entity.type.wirerod;
+package com.pobluesky.backend.domain.lineitem.entity.type;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -9,7 +9,10 @@ public enum InqName {
 
     JS_SI123("JS-SI123"),
     JS_SI456("JS-SI456"),
-    JS_SI789("JS-SI789");
+    JS_SI789("JS-SI789"),
+    JS_SI321("JS-SI321"),
+    JS_SI654("JS-SI654"),
+    JS_SI987("JS-SI987");
 
     private String name;
 }
